Move adding `--internet-tests` option back to `conftest.py`
It needs to be where `pytest` gets called from, otherwise it doesn't work.
