+++ scanpy/tests/conftest.py
@@ -1,34 +1,14 @@
 from pathlib import Path
 
-import pytest
-
 import matplotlib as mpl
 mpl.use('agg')
 from matplotlib import pyplot
 from matplotlib.testing.compare import compare_images
+import pytest
 
 import scanpy
 
 scanpy.settings.verbosity = "hint"
-
-
-def pytest_addoption(parser):
-    parser.addoption(
-        "--internet-tests",
-        action="store_true",
-        default=False,
-        help="Run tests that retrieve stuff from the internet. This increases test time.",
-    )
-
-
-def pytest_collection_modifyitems(config, items):
-    run_internet = config.getoption("--internet-tests")
-    skip_internet = pytest.mark.skip(reason="need --internet-tests option to run")
-    for item in items:
-        # All tests marked with `pytest.mark.internet` get skipped unless
-        # `--run-internet` passed
-        if not run_internet and ("internet" in item.keywords):
-            item.add_marker(skip_internet)
 
 
 def make_comparer(path_expected: Path, path_actual: Path, *, tol: int):
